string module and more tests
